Fixed requires for simplecov
