+++ lib/structured_store.rb
@@ -2,6 +2,10 @@ require 'structured_store/version'
 
 require 'zeitwerk'
 loader = Zeitwerk::Loader.for_gem
+
+# Avoid loading default resolvers by default
+loader.ignore("#{__dir__}/structured_store/ref_resolvers/defaults.rb")
+
 loader.setup
 
 require 'structured_store/engine'
