+++ lib/structured_store/ref_resolvers/defaults.rb
@@ -1,0 +1,5 @@
+# This file will require the registry and default resolvers.
+require 'structured_store/ref_resolvers/registry'
+
+require 'structured_store/ref_resolvers/blank_ref_resolver'
+require 'structured_store/ref_resolvers/definitions_resolver'
